chore: remove repeated code: generateMetaTags
